part4d blog expansion step9

+++ part4/blog/server/controllers/blogs.js
@@ -4,6 +4,17 @@ const User = require('../models/user')
 const jwt = require('jsonwebtoken')
 
 
+const retrieveUser = async (request, response) => {
+  // Authenticating User before adding blog 
+  const decodedToken = jwt.verify(request.token, process.env.SECRET)
+  if (!decodedToken.id) {
+    return response.status(401).json({ error: 'token missing or invalid'})
+  }
+  const user = await User.findById(decodedToken.id)
+  console.log(user)
+  return user
+}
+
 blogRouter.get('/', async (request, response) => {
   const blog = await Blog
     .find({}).populate('user', { username: 1, name: 1 })
@@ -14,13 +25,7 @@ blogRouter.post('/', async (request, response) => {
   const body = request.body
 
   // Authenticating User before adding blog 
-  console.log(request.token)
-  const decodedToken = jwt.verify(request.token, process.env.SECRET)
-  if (!decodedToken.id) {
-    return response.status(401).json({ error: 'token missing or invalid'})
-  }
-  const user = await User.findById(decodedToken.id)
-
+  const user = await retrieveUser(request, response)
 
   if (body.title === undefined || body.url === undefined) {
     response.status(400).end()
@@ -44,8 +49,17 @@ blogRouter.post('/', async (request, response) => {
 })
 
 blogRouter.delete('/:id', async (request, response) => {
-  await Blog.findByIdAndRemove(request.params.id)
-  response.status(204).end()
+
+  const user = await retrieveUser(request, response)
+  const blog = await Blog.findById(request.params.id)
+
+  if(blog.user.toString() === user.id) {
+    await Blog.findByIdAndRemove(request.params.id)
+    response.status(204).end()
+  } else {
+    return response.status(401).json({ error: 'token missing or invalid'})
+  }
+
 })
